+++ src/io/+mag/+io/+out/Format.m
@@ -1,5 +1,5 @@
 classdef (Abstract) Format < matlab.mixin.Heterogeneous & mag.mixin.SetGet
-% IFORMAT Interface for data format providers for export.
+% FORMAT Interface for data format providers for export.
 
     properties (Abstract, Constant)
         % EXTENSION Extension supported for file format.
@@ -11,8 +11,8 @@ classdef (Abstract) Format < matlab.mixin.Heterogeneous & mag.mixin.SetGet
         % GETEXPORTFILENAME Get name of export file name.
         fileName = getExportFileName(this, data)
 
-        % CONVERTTOEXPORTABLEFORMAT Convert data to an exportable format.
-        exportData = convertToExportableFormat(this, data)
+        % CONVERTTOEXPORTFORMAT Convert data to an exportable format.
+        exportData = convertToExportFormat(this, data)
 
         % WRITE Export file to format.
         write(this, fileName, exportData)